Add pure stock and flow integration engine
Version 1.x of the package requires to specify the box model in a graph
format that includes string identifiers for all elements as well as
evalutators for all flows and variables. In addition, there are
parameters that are intended to be modyfied by the user during runtime.
This format puts a lot of restrictions on the engine implementation,
e.g. it requires traversal of the implicit evaluation graph during each
run without the possibility for caching since the user-supplied
evalutators could theoretically define different implicit graphs for
different runs. Accessing graph elements via string identifiers slows
down the integration even further.

This patch is the first attempt to factor out the engine such that it
operates only on stocks (`number[]`) and flows
(`{ value: number; from: number; to: number }[]`). The user has to
supply a function that computes the flows for a given array of stock
values (and timestamp, if required).

This function could directly compute all elements of a specific box
model very efficiently without having to explicitly define the graph.
On the other hand, it could utilize the previous (still to
refactor) graph evaluator for compatibility reasons or define its own
evaluator for yet another box model format.